change scroll assets
