+++ Mac_Health/Assets.xcassets/Influencer/RotateImage/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
