+++ Mac_Health/Assets.xcassets/Influencer/RotateImage/Influencer1/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
