+++ Mac_Health/Common/Component/TopImage.swift
@@ -20,7 +20,7 @@ struct TopImage: View {
                     SubscribeView()
                         .navigationBarTitle("정회승의 Smart Routine", displayMode: .inline)
                 } label: {
-                    RecommendPage()
+                    RecommendPage(RecommendBannerNum: idx)
                 }
             }
         }
@@ -40,7 +40,41 @@ struct TopImage: View {
         .frame(height: UIScreen.getHeight(358))
         .tabViewStyle(.page)
     }
+    
 }
+
+struct RecommendPage: View {
+    
+    var RecommendBannerNum: Int
+    
+    var body: some View {
+        
+        ZStack(alignment: .bottomTrailing){
+            Color.gray_900.ignoresSafeArea()
+            
+            Image("RecommendBanner\(RecommendBannerNum)")
+                .resizable()
+                .frame(width: UIScreen.getWidth(390), height: UIScreen.getHeight(358))
+            
+            LinearGradient(colors: [.gray_900,.clear, .clear, .gray_900.opacity(0.7), .gray_900], startPoint: .top, endPoint: .bottom)
+            HStack{
+                VStack(alignment: .leading, spacing: 16){
+                    Text("정회승의 Smart Routine")
+                        .foregroundColor(.label_900)
+                        .font(.title1())
+                    Text("2022 Mr. 서울대🏆")
+                        .foregroundColor(.label_600)
+                        .font(.body2())
+                        .padding(.bottom, 40)
+                }
+                Spacer()
+            }
+            .padding(.leading, 20)
+        }
+        .frame(height: UIScreen.getHeight(358))
+    }
+}
+
 
 
 struct TopImageView_Previews: PreviewProvider {
